+++ nova/modules/core/language/english/text_lang.php
@@ -181,6 +181,7 @@ $lang['mark_as_read'] = "Mark All as Read";
 $lang['versions_redirect'] = "For a full version history of Nova, please visit <a href='http://docs.anodyne-productions.com/index.php/nova2/overview/index' target='_blank'>AnodyneDocs</a>.";
 $lang['components_redirect'] = "For a full listing of third-party components and plugins in Nova, please visit <a href='http://docs.anodyne-productions.com/index.php/nova2/overview/index' target='_blank'>AnodyneDocs</a>.";
 $lang['tags_explain'] = "Tags (i.e. keywords) are a form of meta-data to classify items and help find them later using Nova's search capabilities (e.g. klingon, peace summit, Khitomer)";
+$lang['metadata_explain'] = "You can specify additional information to be displayed on the manifest by entering the field's <strong>HTML name</strong> (found under HTML Attributes when editing a bio form field) and separating fields with a pipe (|). Information will be displayed in the order specified. (i.e. species|gender)";
 
 /*
 |---------------------------------------------------------------